Fix CI: unnecessary cast, non-canonical fixture paths, symlink test, MSRV
clippy (Linux/macOS)
  MetadataExt::dev returns u64 on every Unix target, so the cast in
  platform::unix::volume_id was redundant. Windows never compiles that file,
  which is why only two of the three clippy jobs caught it.

test (Windows/macOS)
  The scanner canonicalises its root, so a test that compares against the raw
  tempdir path is comparing two different spellings of the same place. That is
  invisible where the temp directory is already canonical, and breaks where it
  is not: macOS reaches /var/folders through a symlink to /private/var, and
  Windows CI runners expose the temp directory under the 8.3 short name
  RUNNER~1. Fixture now canonicalises its root once, so every path a test
  builds matches what the scanner reports.

test (Linux/macOS)
  symlinked_directories_are_not_followed_by_default asserted an exact byte
  total, but the walker counts the link itself, and a symlink's size is the
  length of the path it stores. The test now asserts the property it meant to
  check: the target is not counted twice, and the link does not become a node
  in the tree. These cases only run under cfg(unix), so they had never
  executed on the machine this was written on.

minimum supported Rust version
  The job deleted Cargo.lock before checking, which resolved whatever the
  registry offers today and so measured our dependencies rather than us. It
  pulled a crate requiring edition 2024 and could not have passed at 1.82. It
  now checks --locked against the committed lockfile, which is the dependency
  set users actually get.

  The real minimum is 1.85 for the CLI and 1.86 for the desktop app, both
  verified against real 1.85.0 and 1.86.0 toolchains rather than assumed.
  Cargo's MSRV-aware resolver keeps future lockfile updates honest.

  The screenshot harness moves behind a 'screenshots' feature. Its PNG encoder
  was dragging the whole image crate into the shipped app and pushing that
  binary's minimum to 1.88; it now uses the much smaller png crate. The
  harness also bounds its waits, so it can no longer hang on an event that
  will never arrive.

clippy now runs on all three platforms as well, since the cfg-gated platform
code differs per OS and a Linux-only lint pass never sees two thirds of it.

+++ src/platform/unix.rs
@@ -45,7 +45,9 @@ pub fn is_reparse_point(meta: &Metadata) -> bool {
 }
 
 /// The device number identifying the filesystem an entry lives on.
+///
+/// `MetadataExt::dev` returns `u64` on every Unix target, so no cast is needed.
 #[must_use]
 pub fn volume_id(_path: &Path, meta: &Metadata) -> Option<u64> {
-    Some(meta.dev() as u64)
+    Some(meta.dev())
 }
